requirements.txt
boto3==1.34.112
# Core dependencies
aiohttp==3.9.1
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.1
agentops==0.4.16

# Additional dependencies
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
beautifulsoup4==4.12.2
scikit-learn==1.3.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.1

# aws-bedrock-agent-runtime==0.1.8 (commented out due to unavailability)
